+++ client-old/src/types.ts
@@ -43,8 +43,3 @@ export interface CategoryDetails {
   streaming_required: boolean;
   goals: string[];
 }
-
-export interface Filter {
-  filter: string;
-  value: string | number;
-}
