+++ client-old/src/vite-env.d.ts
