--- client-vue/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <link rel="icon" type="image/svg+xml" href="/favicon.jpg" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>PM64R Statistics</title>
-    </head>
-    <body>
-        <div id="app"></div>
-        <script type="module" src="/src/main.ts"></script>
-    </body>
-</html>
