--- client-vue/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "client-vue",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vue-tsc -b && vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@primeuix/themes": "^1.2.5",
-    "@tailwindcss/vite": "^4.1.14",
-    "@tanstack/vue-query": "^5.90.2",
-    "date-fns": "^4.1.0",
-    "iso8601-duration": "^2.1.3",
-    "primevue": "^4.4.0",
-    "tailwindcss": "^4.1.14",
-    "tailwindcss-primeui": "^0.6.1",
-    "vue": "^3.5.22"
-  },
-  "devDependencies": {
-    "@types/node": "^24.6.0",
-    "@vitejs/plugin-vue": "^6.0.1",
-    "@vue/tsconfig": "^0.8.1",
-    "typescript": "~5.9.3",
-    "vite": "^7.1.7",
-    "vue-tsc": "^3.1.0"
-  }
-}
