--- client-vue/src/main.ts
@@ -1,61 +1,0 @@
-import { createApp } from "vue";
-import { VueQueryPlugin } from "@tanstack/vue-query";
-import PrimeVue from "primevue/config";
-import Aura from "@primeuix/themes/aura";
-import Dropdown from "primevue/dropdown";
-import DataTable from "primevue/datatable";
-import Column from "primevue/column";
-import ProgressSpinner from "primevue/progressspinner";
-import "./style.css";
-import App from "./App.vue";
-
-const app = createApp(App);
-
-// Configure Vue Query
-app.use(VueQueryPlugin, {
-  queryClientConfig: {
-    defaultOptions: {
-      queries: {
-        staleTime: 1000 * 60 * 5, // 5 minutes
-        gcTime: 1000 * 60 * 10, // 10 minutes
-        retry: (failureCount, error) => {
-          // Don't retry on 4xx errors
-          if (
-            error instanceof Error &&
-            "status" in error &&
-            typeof error.status === "number"
-          ) {
-            if (error.status >= 400 && error.status < 500) {
-              return false;
-            }
-          }
-          return failureCount < 3;
-        },
-        refetchOnWindowFocus: false,
-      },
-      mutations: {
-        retry: 1,
-      },
-    },
-  },
-});
-
-// Configure PrimeVue
-app.use(PrimeVue, {
-  theme: {
-    preset: Aura,
-    options: {
-      prefix: "p",
-      darkModeSelector: ".p-dark",
-      cssLayer: false,
-    },
-  },
-});
-
-// Register PrimeVue components globally
-app.component("Dropdown", Dropdown);
-app.component("DataTable", DataTable);
-app.component("Column", Column);
-app.component("ProgressSpinner", ProgressSpinner);
-
-app.mount("#app");
