--- client-vue/src/style.css
@@ -1,216 +1,0 @@
-@import "tailwindcss";
-@import "tailwindcss-primeui";
-
-@font-face {
-    font-family: "PaperMarioFont";
-    src:
-        local("PaperMarioFont"),
-        url("/PaperMarioFont.ttf") format("truetype");
-}
-
-:root {
-    font-family:
-        "PaperMarioFont", system-ui, Avenir, Helvetica, Arial, sans-serif;
-    line-height: 1.5;
-    font-weight: 400;
-
-    color-scheme: light dark;
-    color: rgba(255, 255, 255, 0.87);
-    background-color: #3283b4;
-
-    font-synthesis: none;
-    text-rendering: optimizeLegibility;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-}
-
-* {
-    box-sizing: border-box;
-}
-
-body {
-    margin: 0;
-    padding: 0;
-    min-height: 100vh;
-    background-color: #3283b4;
-}
-
-#app {
-    min-height: 100vh;
-}
-
-/* PrimeVue Dropdown Styling */
-.p-dropdown {
-    background-color: white;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    color: #333;
-}
-
-.p-dropdown:not(.p-disabled):hover {
-    border-color: #3283b4;
-}
-
-.p-dropdown:not(.p-disabled).p-focus {
-    outline: 0 none;
-    outline-offset: 0;
-    box-shadow: 0 0 0 0.2rem rgba(50, 131, 180, 0.2);
-    border-color: #3283b4;
-}
-
-.p-dropdown .p-dropdown-label {
-    color: #333;
-    padding: 0.75rem;
-}
-
-.p-dropdown .p-dropdown-trigger {
-    color: #666;
-}
-
-.p-dropdown-panel {
-    background-color: white;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    box-shadow: 0 2px 12px 0 rgba(0, 0, 0, 0.1);
-}
-
-.p-dropdown-panel .p-dropdown-items .p-dropdown-item {
-    color: #333;
-    padding: 0.75rem;
-    border: 0 none;
-}
-
-.p-dropdown-panel .p-dropdown-items .p-dropdown-item:hover {
-    background-color: #f4f4f4;
-}
-
-.p-dropdown-panel .p-dropdown-items .p-dropdown-item:focus {
-    outline: 0 none;
-    outline-offset: 0;
-    box-shadow: inset 0 0 0 0.15rem rgba(50, 131, 180, 0.2);
-    background-color: #f4f4f4;
-}
-
-.p-dropdown-panel .p-dropdown-items .p-dropdown-item.p-highlight {
-    background-color: #3283b4;
-    color: white;
-}
-
-/* Loading Spinner */
-.p-progress-spinner {
-    color: white;
-}
-
-.p-progress-spinner .p-progress-spinner-circle {
-    stroke: currentColor;
-}
-
-/* DataTable Styling */
-.p-datatable .p-datatable-table {
-    border-collapse: collapse;
-    width: 100%;
-}
-
-.p-datatable .p-datatable-thead > tr > th {
-    text-align: left;
-    padding: 1rem;
-    font-weight: 600;
-    color: white;
-    background-color: rgb(30, 49, 64);
-    border: 1px solid white;
-}
-
-.p-datatable .p-datatable-tbody > tr > td {
-    text-align: left;
-    border: 1px solid #555;
-    padding: 1rem;
-    background-color: rgb(30, 49, 64);
-    color: white;
-}
-
-.p-datatable .p-datatable-tbody > tr:nth-child(even) > td {
-    background-color: rgb(25, 44, 59);
-}
-
-.p-datatable .p-datatable-tbody > tr:hover > td {
-    background-color: rgb(35, 54, 69);
-}
-
-/* Paginator */
-.p-paginator {
-    background-color: rgb(30, 49, 64);
-    color: white;
-    border: none;
-    padding: 1rem;
-    justify-content: center;
-}
-
-.p-paginator .p-paginator-first,
-.p-paginator .p-paginator-prev,
-.p-paginator .p-paginator-next,
-.p-paginator .p-paginator-last {
-    background-color: transparent;
-    border: 1px solid white;
-    color: white;
-    min-width: 3rem;
-    height: 3rem;
-    margin: 0 0.125rem;
-    border-radius: 4px;
-}
-
-.p-paginator .p-paginator-first:not(.p-disabled):hover,
-.p-paginator .p-paginator-prev:not(.p-disabled):hover,
-.p-paginator .p-paginator-next:not(.p-disabled):hover,
-.p-paginator .p-paginator-last:not(.p-disabled):hover {
-    background-color: rgba(255, 255, 255, 0.1);
-}
-
-.p-paginator .p-paginator-pages .p-paginator-page {
-    background-color: transparent;
-    border: 1px solid white;
-    color: white;
-    min-width: 3rem;
-    height: 3rem;
-    margin: 0 0.125rem;
-    border-radius: 4px;
-}
-
-.p-paginator .p-paginator-pages .p-paginator-page:hover {
-    background-color: rgba(255, 255, 255, 0.1);
-}
-
-.p-paginator .p-paginator-pages .p-paginator-page.p-highlight {
-    background-color: #3283b4;
-    border-color: #3283b4;
-    color: white;
-}
-
-.p-paginator .p-dropdown {
-    margin-left: 0.5rem;
-    margin-right: 0.5rem;
-}
-
-/* Custom Paper Mario colors */
-.bg-pm-blue {
-    background-color: #3283b4;
-}
-
-.bg-pm-dark {
-    background-color: rgb(30, 49, 64);
-}
-
-/* Links */
-a {
-    color: #646cff;
-    text-decoration: none;
-}
-
-a:hover {
-    color: #535bf2;
-    text-decoration: underline;
-}
-
-/* Utility classes */
-.hide {
-    display: none !important;
-}
