+++ client/src/components/RawDataTable.vue
