+++ client/src/main.ts
@@ -1,19 +1,61 @@
+import { createApp } from "vue";
+import { VueQueryPlugin } from "@tanstack/vue-query";
+import PrimeVue from "primevue/config";
+import Aura from "@primeuix/themes/aura";
+import Dropdown from "primevue/dropdown";
+import DataTable from "primevue/datatable";
+import Column from "primevue/column";
+import ProgressSpinner from "primevue/progressspinner";
 import "./style.css";
-import {
-  setupUserFilter,
-  setupCategoryFilter,
-  setupRaceTypeFilter,
-} from "./ui-managers/filter.ui.ts";
-import { updateStatistics } from "./coordinators/statistic.coordinator.ts";
+import App from "./App.vue";
 
-setupUserFilter(document.querySelector<HTMLSelectElement>("#user-selector")!);
+const app = createApp(App);
 
-setupCategoryFilter(
-  document.querySelector<HTMLSelectElement>("#category-selector")!
-);
+// Configure Vue Query
+app.use(VueQueryPlugin, {
+  queryClientConfig: {
+    defaultOptions: {
+      queries: {
+        staleTime: 1000 * 60 * 5, // 5 minutes
+        gcTime: 1000 * 60 * 10, // 10 minutes
+        retry: (failureCount, error) => {
+          // Don't retry on 4xx errors
+          if (
+            error instanceof Error &&
+            "status" in error &&
+            typeof error.status === "number"
+          ) {
+            if (error.status >= 400 && error.status < 500) {
+              return false;
+            }
+          }
+          return failureCount < 3;
+        },
+        refetchOnWindowFocus: false,
+      },
+      mutations: {
+        retry: 1,
+      },
+    },
+  },
+});
 
-setupRaceTypeFilter(
-  document.querySelector<HTMLSelectElement>("#race-type-selector")!
-);
+// Configure PrimeVue
+app.use(PrimeVue, {
+  theme: {
+    preset: Aura,
+    options: {
+      prefix: "p",
+      darkModeSelector: ".p-dark",
+      cssLayer: false,
+    },
+  },
+});
 
-updateStatistics();
+// Register PrimeVue components globally
+app.component("Dropdown", Dropdown);
+app.component("DataTable", DataTable);
+app.component("Column", Column);
+app.component("ProgressSpinner", ProgressSpinner);
+
+app.mount("#app");
