+++ client/src/services/categoriesApi.ts
