+++ client/src/services/statisticsApi.ts
